Improving cross platform support with MacOS

+++ includes/renderer/Renderer.h
@@ -22,7 +22,7 @@ class SortRenderer
 public:
     void update(std::vector<int>& elems, int a, int b);
 
-    void renderText(std::string txt, int x, int y, SDL_Color color);
+    void renderText(std::string txt, float x, float y, SDL_Color color);
     void renderInfo();
 
     SDL_Color HSVToRGB(unsigned char hue, unsigned char sat, unsigned char value);
